chore: refactor client
