+++ package.json
@@ -48,7 +48,6 @@
     "it-length-prefixed": "^3.1.0",
     "it-pipe": "^1.1.0",
     "libp2p": "libp2p/js-libp2p#0.30.x",
-    "libp2p-interfaces": "^0.5.1",
     "libp2p-mplex": "^0.10.0",
     "libp2p-noise": "^2.0.1",
     "libp2p-tcp": "^0.15.1",
