+++ src/errors.js
@@ -2,8 +2,5 @@
 
 exports.codes = {
   INVALID_NAMESPACE: 'ERR_INVALID_NAMESPACE',
-  INVALID_TTL: 'ERR_INVALID_TTL',
-  INVALID_LIMIT: 'ERR_INVALID_LIMIT',
-  NO_CONNECTED_RENDEZVOUS_SERVERS: 'ERR_NO_CONNECTED_RENDEZVOUS_SERVERS',
-  INVALID_COOKIE: 'ERR_INVALID_COOKIE'
+  NO_CONNECTED_RENDEZVOUS_SERVERS: 'ERR_NO_CONNECTED_RENDEZVOUS_SERVERS'
 }
