+++ src/server/errors.js
@@ -1,0 +1,5 @@
+'use strict'
+
+exports.codes = {
+  INVALID_COOKIE: 'ERR_INVALID_COOKIE'
+}
